refactor(button): move buttons to separate files, move global styles to component
