+++ src/components/pageComponents/home/Examples/demos/TransactionButtonDemo/NativeTokenDemo.tsx
@@ -1,8 +1,8 @@
 import Wrapper from '@/src/components/pageComponents/home/Examples/demos/TransactionButtonDemo/Wrapper'
 import TransactionButton from '@/src/components/sharedComponents/TransactionButton'
 import { withWalletStatusVerifier } from '@/src/components/sharedComponents/WalletStatusVerifier'
-import { PrimaryButton } from '@/src/components/sharedComponents/ui/Buttons'
 import { GeneralMessage } from '@/src/components/sharedComponents/ui/GeneralMessage'
+import PrimaryButton from '@/src/components/sharedComponents/ui/PrimaryButton'
 import { useWeb3StatusConnected } from '@/src/hooks/useWeb3Status'
 import { Dialog, Portal } from '@chakra-ui/react'
 import { type ReactElement, useState } from 'react'
