+++ src/components/sharedComponents/TransactionButton.tsx
@@ -1,5 +1,5 @@
 import { withWalletStatusVerifier } from '@/src/components/sharedComponents/WalletStatusVerifier'
-import { PrimaryButton } from '@/src/components/sharedComponents/ui/Buttons'
+import PrimaryButton from '@/src/components/sharedComponents/ui/PrimaryButton'
 import { useTransactionNotification } from '@/src/lib/toast/TransactionNotificationProvider'
 import type { ButtonProps } from '@chakra-ui/react'
 import { useEffect, useState } from 'react'
