+++ src/components/sharedComponents/ui/PrimaryButton/styles.ts
@@ -1,0 +1,26 @@
+export const styles = {
+  'html.light &': {
+    '--theme-button-primary-background-color': '#692581',
+    '--theme-button-primary-background-color-hover': '#892fa9',
+    '--theme-button-primary-border-color': '#692581',
+    '--theme-button-primary-border-color-hover': '#892fa9',
+    '--theme-button-primary-color': '#fff',
+    '--theme-button-primary-color-hover': '#fff',
+    '--theme-button-primary-background-color-disabled': '#692581',
+    '--theme-button-primary-border-color-disabled': '#692581',
+    '--theme-button-primary-color-disabled': '#fff',
+  },
+  'html.dark &': {
+    '--theme-button-primary-background-color': '#8b46a4',
+    '--theme-button-primary-background-color-hover': '#9a4eb5',
+    '--theme-button-primary-border-color': '#8b46a4',
+    '--theme-button-primary-border-color-hover': '#9a4eb5',
+    '--theme-button-primary-color': '#fff',
+    '--theme-button-primary-color-hover': '#fff',
+    '--theme-button-primary-background-color-disabled': '#8b46a4',
+    '--theme-button-primary-border-color-disabled': '#8b46a4',
+    '--theme-button-primary-color-disabled': '#fff',
+  },
+}
+
+export default styles
