+++ src/components/sharedComponents/ui/SecondaryButton/styles.ts
@@ -1,0 +1,26 @@
+export const styles = {
+  'html.light &': {
+    '--theme-button-secondary-background-color': '#2e3048',
+    '--theme-button-secondary-background-color-hover': '#3d405f',
+    '--theme-button-secondary-border-color': '#2e3048',
+    '--theme-button-secondary-border-color-hover': '#3d405f',
+    '--theme-button-secondary-color': '#fff',
+    '--theme-button-secondary-color-hover': '#fff',
+    '--theme-button-secondary-background-color-disabled': '#2e3048',
+    '--theme-button-secondary-border-color-disabled': '#2e3048',
+    '--theme-button-secondary-color-disabled': '#fff',
+  },
+  'html.dark &': {
+    '--theme-button-secondary-background-color': '#5f6178',
+    '--theme-button-secondary-background-color-hover': '#4a4c5f',
+    '--theme-button-secondary-border-color': '#5f6178',
+    '--theme-button-secondary-border-color-hover': '#4a4c5f',
+    '--theme-button-secondary-color': '#fff',
+    '--theme-button-secondary-color-hover': '#fff',
+    '--theme-button-secondary-background-color-disabled': '#5f6178',
+    '--theme-button-secondary-border-color-disabled': '#5f6178',
+    '--theme-button-secondary-color-disabled': '#fff',
+  },
+}
+
+export default styles
