+++ src/utils/suspenseWrapper.tsx
@@ -1,5 +1,5 @@
-import { PrimaryButton } from '@/src/components/sharedComponents/ui/Buttons'
 import { GeneralMessage } from '@/src/components/sharedComponents/ui/GeneralMessage'
+import PrimaryButton from '@/src/components/sharedComponents/ui/PrimaryButton'
 import { Spinner } from '@chakra-ui/react'
 import { Dialog, Portal } from '@chakra-ui/react'
 import { QueryErrorResetBoundary } from '@tanstack/react-query'
